requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-migrate==4.0.5
flask-sqlalchemy==3.0.5
flask-mail==0.9.1
flask-socketio==5.3.6
flask-limiter==3.5.0
werkzeug==2.3.7
sqlalchemy==2.0.21
python-dotenv==1.0.0
redis==5.0.0
celery==5.3.2
requests==2.31.0
pillow==10.0.1
google-api-python-client==2.103.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dateutil==2.8.2
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
